+++ src/features/bedrijf/bedrijf-details/index.ts
@@ -1,0 +1,2 @@
+export { default as HandelsregisterGegevens } from "./HandelsregisterGegevens.vue";
+export * from "./use-bedrijf-by-identifier";
