+++ src/features/bedrijf/bedrijf-zoeken/index.ts
@@ -1,0 +1,1 @@
+export { default as BedrijfZoeker } from "./BedrijfZoeker.vue";
