+++ src/features/bedrijf/bedrijf-zoeken/use-klant-by-bedrijf-identifier.ts
@@ -1,0 +1,23 @@
+import { ServiceResult } from "@/services";
+import {
+  findKlantByIdentifier,
+  type KlantBedrijfIdentifier,
+} from "@/services/klanten";
+
+export const useKlantByBedrijfIdentifier = (
+  getId: () => KlantBedrijfIdentifier | undefined,
+) => {
+  const getCacheKey = () => {
+    const id = getId();
+    if (!id) return "";
+    return "klant" + JSON.stringify(getId());
+  };
+  const findKlant = () => {
+    const id = getId();
+    if (!id) {
+      throw new Error();
+    }
+    return findKlantByIdentifier(id);
+  };
+  return ServiceResult.fromFetcher(getCacheKey, findKlant);
+};
