+++ src/features/klant/KlantDetails.vue
@@ -17,7 +17,7 @@
 <script lang="ts" setup>
 import { type PropType } from "vue";
 import { Heading as UtrechtHeading } from "@utrecht/component-library-vue";
-import type { Klant } from "./types";
+import type { Klant } from "../../services/klanten/types";
 
 defineProps({
   klant: {
