--- src/features/klant/bedrijf/service/bedrijf-query.ts
@@ -1,5 +1,0 @@
-import type { BedrijfQuery, SearchCategories } from "../types";
-
-export const bedrijfQuery = <K extends SearchCategories>(
-  query: BedrijfQuery<K>,
-) => query;
