+++ src/features/klant/use-klant-by-id.ts
@@ -1,6 +1,6 @@
 import { ServiceResult } from "@/services";
 import type { Ref } from "vue";
-import { fetchKlantById } from "./service";
+import { fetchKlantById } from "../../services/klanten/service";
 
 export const useKlantById = (id: Ref<string>) =>
   ServiceResult.fromFetcher(() => id.value || "", fetchKlantById);
