+++ src/features/persoon/persoon-details/index.ts
@@ -1,0 +1,2 @@
+export { default as BrpGegevens } from "./BrpGegevens.vue";
+export * from "./use-persoon-by-bsn";
