+++ src/features/persoon/persoon-zoeken/ensure-klant-for-bsn.ts
@@ -1,0 +1,11 @@
+import {
+  findKlantByIdentifier,
+  createKlant,
+  type Contactnaam,
+} from "../../../services/klanten/service";
+
+export const ensureKlantForBsn = async (parameters: {
+  bsn: string;
+  contactnaam: Contactnaam;
+}) =>
+  (await findKlantByIdentifier(parameters)) ?? (await createKlant(parameters));
