+++ src/features/persoon/persoon-zoeken/index.ts
@@ -1,0 +1,1 @@
+export { default as PersoonZoeker } from "./PersoonZoeker.vue";
