+++ src/services/kvk/index.ts
@@ -1,0 +1,2 @@
+export * from "./service";
+export * from "./types";
